refactor helper to a base class
